Add support to pull SMBIOS values out of UEFI vars
This patch demonstrates how to retrieve strings from UEFI variables for
use in SMBIOS tables. Both the system serial number and system SKU
number are retrieved from UEFI variables.

The namespace used is {72096f5b-2ac7-4e6d-a7bb-bf947d673415}. This is so
that other UEFI drivers and OS components are able to populate and
retrieve these values as necessary

Note that FakeSMBIOSDataInVolatileStorage is called to aid in testing
until non-volatile UEFI variables are enabled.

Signed-off-by: Henry Beberman <hebeberm@microsoft.com>