Add VirtualGopDxe driver to iMXPlatformPkg (#30)
This patch adds a new VirtualGopDxe driver which advertises a virtual 1920x1080 display to Windows. The user can view the contents of this virtual display remotely in the Windows Device Portal.

In addition, this patch refactors the Gop QueryMode/SetMode/Blt callbacks into the iMXDisplayLib and updates the different GopDxe drivers to use the library functions where appropriate.