Add json parser
This commit adds code to parse a simplified json file. The json file must
be only one level and all key/value pairs must be encapsulated by quotation marks.
We use this json file to populate key/value nodes in a linked list which will
be written to NV store.