Move Optee libs to use toolchain libc
Currently our Optee components depend on edk2-libc, which is now
incompatible with the ArmPkg. To reduce complexity, convert Optee
components to use the toolchain-provided stdlib.

limits.h and string.h ends up including gnu/stubs-soft.h which
is not included in the arm-linux-gnueabihf toolchain. Edk2 expects to
be built with float=soft yet the officially supported toolchain is a
this hardfloat based one. We don't actually use these functions in our
code so just remove the headers.