Fix build break - remove dependency on edk2 stdlib package (#42)
Originally, we always include Edk2's StdLib.inc into our build. This
lib is only meant to be included to build UEFI applications but for some
reason we decided to have it included in all components (PEI, DXE, etc).
Including this file sets the -nostdinc flag, presumably to avoid
collisions with the toolchain's stdlib.

edk2 recently updated their ArmPkg/ArmSoftFloatLib library to use a
modern implementation. This implementation requires stdint.h and
stdbool.h and expects to pull these from the toolchain and not
edk2's stdlib implementation.

We shouldn't be using edk2's StdLib for building non-application
components in the first place so this patch removes our dependency
on it.

Signed-off-by: Christopher Co christopher.co@microsoft.com